+++ src/main/java/zedly/zbot/event/EnterCreditsEvent.java
@@ -2,13 +2,13 @@ package zedly.zbot.event;
 
 public class EnterCreditsEvent extends Event {
 
-    private float value;
+    private double value;
 
-    public EnterCreditsEvent(float value) {
+    public EnterCreditsEvent(double value) {
         this.value = value;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 }
